Make a type alias for our entry format.
This will make it easier to do an upcoming refactoring to provide more
detail on each entry, as a test run with 1Password shows that they
expect there to be at least 3 (and probably 4+) columns, for username,
password, website, and potentially extra notes.

A lot of this information is obtainable but will require more granular
storage of entry data.